Show how to use the modules to organise the store
