+++ src/App.vue
@@ -47,7 +47,7 @@ export default {
   name: 'app',
   computed: {
     cart() {
-      return this.$store.state.cart;
+      return this.$store.state.roborts.cart;
     },
   },
 };
