+++ src/cart/ShoppingCart.vue
@@ -54,7 +54,7 @@ export default {
   name: 'Cart',
   computed: {
     cart() {
-      return this.$store.state.cart;
+      return this.$store.state.roborts.cart;
     },
     cartSaleItems() {
       return this.$store.getters.cartSaleItems;
