+++ src/parts/get-parts-mixin.js
@@ -4,7 +4,7 @@ export default {
   },
   computed: {
     parts() {
-      return this.$store.state.parts || {
+      return this.$store.state.roborts.parts || {
         heads: [],
         arms: [],
         torsos: [],
