+++ src/store/modules/users.js
@@ -1,0 +1,21 @@
+import axios from 'axios';
+
+export default {
+  state: {
+    user: null,
+  },
+  mutations: {
+    updateCurrentUser(state, user) {
+      state.user = user;
+    },
+  },
+  getters: {
+  },
+  actions: {
+    signIn({ commit }) {
+      axios.post('/api/sign-in')
+        .then(result => commit('updateCurrentUser', result.data))
+        .catch(console.error);
+    },
+  },
+};
